chore(arrow) Text Renderer Split
