+++ docs/api-reference/text/README.md
@@ -1,36 +1,35 @@
 # @luma.gl/text
 
-`@luma.gl/text` provides experimental 2D text rendering utilities built around a clean split between data conversion and rendering.
+`@luma.gl/text` provides experimental GPU-only 2D text rendering utilities.
 
-Layer or data-preparation code owns Arrow source data. It converts Arrow vectors into `GPUVector` objects or prepared GPU state, then constructs renderer models that only know about GPU-resident resources.
+`@luma.gl/arrow` owns Arrow source data, source mapping, upload, and Arrow UTF-8/dictionary preparation. It converts Arrow vectors into `GPUVector` objects or prepared GPU state, then constructs renderer models that only know about GPU-resident resources.
 
 ## Public Architecture
 
 | Responsibility | Public APIs |
 | --- | --- |
-| Arrow-to-GPUVector upload | `convertArrowTextToAttribute()`, `convertArrowTextToStorage()`, `convertArrowTextToDictionary()` |
-| Arrow-to-prepared-state conversion | `convertArrowTextToAttributeState()`, `convertArrowTextToStorageState()`, `convertArrowTextToDictionaryState()` |
-| Pure rendering | `AttributeTextModel`, `StorageTextModel`, `RowIndexedStorageTextModel`, `DictionaryTextModel` |
-| UTF-8/glyph utilities | `buildArrowUtf8Chunks()`, `buildGpuUtf8TextInput()`, `buildGpuDictionaryUtf8TextInput()`, `buildArrowGlyphLayout()` |
+| Arrow facade in `@luma.gl/arrow` | `ArrowTextRenderer`, `resolveArrowTextSourceVectors()`, `convertArrowTextToAttribute()`, `convertArrowTextToStorage()`, `convertArrowTextToDictionary()` |
+| Pure rendering in `@luma.gl/text` | `AttributeTextModel`, `StorageTextModel`, `RowIndexedStorageTextModel`, `DictionaryTextModel` |
+| GPU input contracts in `@luma.gl/text` | `ATTRIBUTE_TEXT_GPU_INPUT_SCHEMA`, `STORAGE_TEXT_GPU_INPUT_SCHEMA`, `DICTIONARY_TEXT_GPU_INPUT_SCHEMA` |
 
-The deprecated Arrow wrapper models are intentionally not exported. New code should perform conversion in the layer and construct one of the pure models with prepared state.
+New code should perform Arrow conversion through `@luma.gl/arrow` and construct one of the pure models with flat prepared props.
 
 ## Attribute Path
 
-Use `convertArrowTextToAttribute()` to upload Arrow source vectors, then `convertArrowTextToAttributeState()` to build glyph-instance attributes. Pass the result to `AttributeTextModel`.
+Use `convertArrowTextToAttribute()` to upload Arrow source vectors, then `convertArrowTextToAttributeModelProps()` to build flat model props for `AttributeTextModel`.
 
 The attribute path supports row colors and per-character color lists. It expands text rows into generated glyph vertex attributes and renders through a GPU table.
 
 ## Storage Path
 
-Use `convertArrowTextToStorage()` and `convertArrowTextToStorageState()` when rendering with WebGPU storage buffers. Pass the resulting state to `StorageTextModel` or `RowIndexedStorageTextModel`.
+Use `convertArrowTextToStorage()` and `convertArrowTextToStorageModelProps()` when rendering with WebGPU storage buffers. Pass the resulting flat props to `StorageTextModel` or `RowIndexedStorageTextModel`.
 
 `RowIndexedStorageTextModel` stores one extra source-row index per generated glyph. This avoids shader-side row lookup by binary search at the cost of a larger generated glyph vertex record.
 
 ## Dictionary Path
 
-Use `convertArrowTextToDictionary()` and `convertArrowTextToDictionaryState()` for dictionary-encoded UTF-8 text. The dictionary model stores shared glyph records per dictionary value and per-row dictionary references.
+Use `convertArrowTextToDictionary()` and `convertArrowTextToDictionaryModelProps()` for dictionary-encoded UTF-8 text. The dictionary model stores shared glyph records per dictionary value and per-row dictionary references.
 
 ## Resource Ownership
 
-Conversion results and prepared states own GPU resources. Layers should destroy converted `GPUVector` bundles when they are replaced or removed. Models should be constructed with `ownsAttributeState` or `ownsStorageState` when the model should destroy prepared state resources.
+Conversion results and prepared props own GPU resources. Layers should destroy converted `GPUVector` bundles when they are replaced or removed. Models should be constructed with `ownsAttributeState` or `ownsStorageState` when the model should destroy prepared resources.
