+++ modules/arrow/package.json
@@ -45,6 +45,7 @@
     "@luma.gl/engine": "~9.3.0",
     "@luma.gl/shadertools": "~9.3.0",
     "@luma.gl/tables": "~9.3.0",
+    "@luma.gl/text": "~9.3.0",
     "@math.gl/geoarrow": "workspace:*",
     "@math.gl/polygon": "^4.1.0",
     "@math.gl/types": "^4.1.0",
