+++ modules/arrow/src/arrow/renderers/text/preparation/index.ts
@@ -1,0 +1,75 @@
+// luma.gl
+// SPDX-License-Identifier: MIT
+// Copyright (c) vis.gl contributors
+
+export {
+  buildArrowGlyphLayout,
+  buildArrowUtf8Chunks,
+  buildGpuDictionaryCompressedTextStream,
+  buildGpuDictionaryUtf8TextInput,
+  buildGpuExpandedTextStream,
+  buildGpuUtf8TextInput,
+  createArrowUtf8TextIndexAccessor,
+  decodeArrowUtf8CodePoints,
+  isArrowUtf8DictionaryType,
+  isArrowUtf8DictionaryVector,
+  isArrowUtf8TextVector,
+  isArrowUtf8Vector,
+  populateUtf8TextIndices,
+  type ArrowGlyphLayout,
+  type ArrowUtf8Dictionary,
+  type ArrowUtf8DictionaryIndexType,
+  type ArrowUtf8Chunk,
+  type ArrowUtf8TextAccessorContext,
+  type ArrowUtf8TextIndexAccessor,
+  type ArrowUtf8TextType,
+  type ArrowUtf8TextVector,
+  type Utf8TextIndexTarget
+} from './arrow-text';
+export {
+  buildArrowTextGlyphTable,
+  createArrowAttributeTextState,
+  createArrowDictionaryStorageTextState,
+  createArrowStorageTextState,
+  createStorageTextStateFromGPUVectors,
+  packStorageTextClipRects,
+  type ArrowAttributeTextInputProps,
+  type ArrowAttributeTextRenderProps,
+  type ArrowAttributeTextState,
+  type ArrowDictionaryStorageTextBatchState,
+  type ArrowDictionaryStorageTextInputProps,
+  type ArrowDictionaryStorageTextRenderBatchState,
+  type ArrowDictionaryStorageTextRenderProps,
+  type ArrowDictionaryStorageTextSourceVectors,
+  type ArrowDictionaryStorageTextState,
+  type ArrowStorageTextBatchState,
+  type ArrowStorageTextInputProps,
+  type ArrowStorageTextRenderBatchState,
+  type ArrowStorageTextRenderProps,
+  type ArrowStorageTextSourceVectors,
+  type ArrowStorageTextState,
+  type ArrowTextGlyphTable,
+  type ArrowTextModelProps,
+  type ArrowTextRenderBatchState,
+  type ArrowTextSourceVectors,
+  type GPUVectorStorageTextBatch,
+  type GPUVectorStorageTextInputProps
+} from './convert-arrow-text-vectors';
+export {
+  convertArrowTextToAttribute,
+  convertArrowTextToAttributeModelProps,
+  convertArrowTextToAttributeState,
+  type ArrowTextConversionColumns,
+  type ConvertedArrowTextData,
+  type ConvertArrowTextProps
+} from './convert-arrow-text-to-attribute';
+export {
+  convertArrowTextToStorage,
+  convertArrowTextToStorageModelProps,
+  convertArrowTextToStorageState
+} from './convert-arrow-text-to-storage';
+export {
+  convertArrowTextToDictionary,
+  convertArrowTextToDictionaryModelProps,
+  convertArrowTextToDictionaryState
+} from './convert-arrow-text-to-dictionary';
