+++ modules/arrow/src/arrow/renderers/text/source/index.ts
@@ -1,0 +1,15 @@
+// luma.gl
+// SPDX-License-Identifier: MIT
+// Copyright (c) vis.gl contributors
+
+export {
+  resolveArrowTextSourceVectors,
+  type ArrowTextColumnSelector,
+  type ArrowTextMappedSourceVectors,
+  type ArrowTextSourceData,
+  type ArrowTextSourceVectorSelectors,
+  type CharacterColorDataType,
+  type OptionalArrowTextColumnSelector,
+  type ResolveArrowTextSourceVectorsProps,
+  type RowColorColumnDataType
+} from './arrow-text-source-mapping';
