+++ modules/arrow/src/index.ts
@@ -99,6 +99,130 @@ export {
   type ArrowPathSourceVectorSelectors,
   type ResolveArrowPathSourceVectorsProps
 } from './arrow/renderers/path/source/arrow-path-source-mapping';
+export {
+  buildArrowGlyphLayout,
+  buildArrowUtf8Chunks,
+  buildGpuDictionaryCompressedTextStream,
+  buildGpuDictionaryUtf8TextInput,
+  buildGpuExpandedTextStream,
+  buildGpuUtf8TextInput,
+  createArrowUtf8TextIndexAccessor,
+  decodeArrowUtf8CodePoints,
+  isArrowUtf8DictionaryType,
+  isArrowUtf8DictionaryVector,
+  isArrowUtf8TextVector,
+  isArrowUtf8Vector,
+  populateUtf8TextIndices,
+  type ArrowGlyphLayout,
+  type ArrowUtf8Dictionary,
+  type ArrowUtf8DictionaryIndexType,
+  type ArrowUtf8Chunk,
+  type ArrowUtf8TextAccessorContext,
+  type ArrowUtf8TextIndexAccessor,
+  type ArrowUtf8TextType,
+  type ArrowUtf8TextVector,
+  type Utf8TextIndexTarget
+} from './arrow/renderers/text/preparation/arrow-text';
+export {
+  buildArrowTextGlyphTable,
+  createArrowAttributeTextState,
+  createArrowDictionaryStorageTextState,
+  createArrowStorageTextState,
+  createStorageTextStateFromGPUVectors,
+  packStorageTextClipRects,
+  type ArrowAttributeTextInputProps,
+  type ArrowAttributeTextRenderProps,
+  type ArrowAttributeTextState,
+  type ArrowDictionaryStorageTextBatchState,
+  type ArrowDictionaryStorageTextInputProps,
+  type ArrowDictionaryStorageTextRenderBatchState,
+  type ArrowDictionaryStorageTextRenderProps,
+  type ArrowDictionaryStorageTextSourceVectors,
+  type ArrowDictionaryStorageTextState,
+  type ArrowStorageTextBatchState,
+  type ArrowStorageTextInputProps,
+  type ArrowStorageTextRenderBatchState,
+  type ArrowStorageTextRenderProps,
+  type ArrowStorageTextSourceVectors,
+  type ArrowStorageTextState,
+  type ArrowTextGlyphTable,
+  type ArrowTextModelProps,
+  type ArrowTextRenderBatchState,
+  type ArrowTextSourceVectors,
+  type GPUVectorStorageTextBatch,
+  type GPUVectorStorageTextInputProps
+} from './arrow/renderers/text/preparation/convert-arrow-text-vectors';
+export {
+  convertArrowTextToAttribute,
+  convertArrowTextToAttributeModelProps,
+  convertArrowTextToAttributeState,
+  type ArrowTextConversionColumns,
+  type ConvertedArrowTextData,
+  type ConvertArrowTextProps
+} from './arrow/renderers/text/preparation/convert-arrow-text-to-attribute';
+export {
+  convertArrowTextToStorage,
+  convertArrowTextToStorageModelProps,
+  convertArrowTextToStorageState
+} from './arrow/renderers/text/preparation/convert-arrow-text-to-storage';
+export {
+  convertArrowTextToDictionary,
+  convertArrowTextToDictionaryModelProps,
+  convertArrowTextToDictionaryState
+} from './arrow/renderers/text/preparation/convert-arrow-text-to-dictionary';
+export {
+  ArrowTextRenderer,
+  addArrowTextGPUTableBatch,
+  createArrowTextGPUTable,
+  createArrowTextGPUTableFromTable,
+  prepareArrowTextInput,
+  prepareArrowTextInputFromData,
+  prepareArrowTextInputFromGPUTable,
+  type ArrowTextRendererActiveModel,
+  type ArrowTextRendererData,
+  type ArrowTextRendererDataBatchUpdate,
+  type ArrowTextRendererInput,
+  type ArrowTextRendererPrepareDataProps,
+  type ArrowTextRendererPrepareGPUTableDataProps,
+  type ArrowTextRendererPrepareInputProps,
+  type ArrowTextRendererProps,
+  type ArrowTextRendererSetPropsResult,
+  type ArrowTextRendererSource,
+  type CharacterColorDataType,
+  type RowColorColumnDataType
+} from './arrow/renderers/text/renderers/arrow-text-renderer';
+export {
+  resolveArrowTextSourceVectors,
+  type ArrowTextColumnSelector,
+  type ArrowTextMappedSourceVectors,
+  type ArrowTextSourceData,
+  type ArrowTextSourceVectorSelectors,
+  type OptionalArrowTextColumnSelector,
+  type ResolveArrowTextSourceVectorsProps
+} from './arrow/renderers/text/source/index';
+export {
+  createArrowTextPickingManager,
+  createArrowTextPickingModel,
+  drawArrowTextPickingPass,
+  getArrowTextRenderModules,
+  supportsTextIndexPicking
+} from './arrow/renderers/text/renderers/arrow-text-picking';
+export {
+  createArrowTextShaderInputs,
+  type ArrowTextShaderInputs
+} from './arrow/renderers/text/renderers/arrow-text-shaders';
+export {
+  AttributeTextModel,
+  DictionaryTextModel,
+  RowIndexedStorageTextModel,
+  StorageTextModel,
+  ATTRIBUTE_TEXT_GPU_INPUT_SCHEMA,
+  DICTIONARY_TEXT_GPU_INPUT_SCHEMA,
+  STORAGE_TEXT_GPU_INPUT_SCHEMA,
+  type AttributeTextModelProps,
+  type DictionaryTextModelProps,
+  type StorageTextModelProps
+} from '@luma.gl/text';
 
 export {
   makeGPUDataFromArrowData,
