+++ modules/arrow/test/arrow/arrow-renderer-preparation.node.spec.ts
@@ -5,6 +5,8 @@
 import test from '@luma.gl/devtools-extensions/tape-test-utils';
 import {
   ArrowPolygonRenderer,
+  addArrowTextGPUTableBatch,
+  createArrowTextGPUTable,
   makeArrowFixedSizeListVector,
   prepareArrowPolygonInput,
   resolveArrowPickInfo
@@ -22,10 +24,6 @@ import {
   ArrowPointRenderer,
   prepareArrowPointInput
 } from '../../../../examples/arrow/arrow-points/arrow-point-renderer';
-import {
-  addArrowTextGPUTableBatch,
-  createArrowTextGPUTable
-} from '../../../../examples/arrow/arrow-text-2d/arrow-text-renderer';
 
 type PathArrowType = arrow.List<arrow.FixedSizeList<arrow.Float32>>;
 
@@ -268,6 +266,32 @@ test('ArrowTextRenderer streaming GPU table retains pick source metadata', t => 
     resolveArrowPickInfo({batchIndex: 1, objectIndex: 2}, gpuTable),
     {batchIndex: 1, rowIndex: 2, batchRowIndex: 0},
     'resolves text pick info to a source batch row'
+  );
+
+  gpuTable.destroy();
+  t.end();
+});
+
+test('ArrowTextRenderer GPU table maps nested source selectors into text input names', t => {
+  const device = new NullDevice({});
+  const sourceVectors = makeTextSourceVectors();
+  const recordBatch = makeNestedTextRecordBatch('source', sourceVectors);
+  const gpuTable = createArrowTextGPUTable(device, recordBatch, {
+    positions: 'source.positions',
+    texts: 'source.texts',
+    pixelOffsets: 'source.pixelOffsets',
+    textAnchors: 'source.textAnchors',
+    alignmentBaselines: 'source.alignmentBaselines'
+  });
+
+  t.equal(gpuTable.gpuVectors.positions?.format, 'float32x2', 'maps nested positions');
+  t.equal(gpuTable.gpuVectors.texts?.format, 'value-list<uint8>', 'maps nested texts');
+  t.equal(gpuTable.gpuVectors.pixelOffsets?.format, 'float32x2', 'maps nested pixel offsets');
+  t.equal(gpuTable.gpuVectors.textAnchors?.format, 'uint8', 'maps nested text anchors');
+  t.equal(
+    gpuTable.gpuVectors.alignmentBaselines?.format,
+    'uint8',
+    'maps nested alignment baselines'
   );
 
   gpuTable.destroy();
@@ -521,6 +545,44 @@ function makeTextRecordBatch(positions: Float32Array, texts: string[]): arrow.Re
   }).batches[0];
   if (!recordBatch) {
     throw new Error('Expected Arrow table to contain a record batch');
+  }
+  return recordBatch;
+}
+
+function makeTextSourceVectors() {
+  return {
+    positions: makeArrowFixedSizeListVector(new arrow.Float32(), 2, new Float32Array([0, 0, 1, 1])),
+    texts: arrow.vectorFromArray(['alpha', 'beta'], new arrow.Utf8()),
+    pixelOffsets: makeArrowFixedSizeListVector(
+      new arrow.Float32(),
+      2,
+      new Float32Array([1, 2, 3, 4])
+    ),
+    textAnchors: arrow.vectorFromArray([0, 1], new arrow.Uint8()),
+    alignmentBaselines: arrow.vectorFromArray([0, 2], new arrow.Uint8())
+  };
+}
+
+function makeNestedTextRecordBatch(
+  fieldName: string,
+  sourceVectors: ReturnType<typeof makeTextSourceVectors>
+): arrow.RecordBatch {
+  const table = new arrow.Table(sourceVectors);
+  const innerStructData = table.batches[0]?.data;
+  if (!innerStructData) {
+    throw new Error('Expected Arrow table to contain a source batch');
+  }
+  const schema = new arrow.Schema([new arrow.Field(fieldName, innerStructData.type)]);
+  const structData = arrow.makeData({
+    type: new arrow.Struct(schema.fields),
+    length: table.numRows,
+    nullCount: 0,
+    nullBitmap: null,
+    children: [innerStructData]
+  });
+  const recordBatch = new arrow.Table([new arrow.RecordBatch(schema, structData)]).batches[0];
+  if (!recordBatch) {
+    throw new Error('Expected nested Arrow table to contain a record batch');
   }
   return recordBatch;
 }
