+++ modules/text/src/text-2d/index.ts
@@ -21,39 +21,11 @@ export {
   type CharacterMapping
 } from './atlas/text-utils';
 export {
-  buildArrowGlyphLayout,
-  buildArrowUtf8Chunks,
-  buildGpuDictionaryCompressedTextStream,
-  buildGpuDictionaryUtf8TextInput,
-  buildGpuExpandedTextStream,
-  buildGpuUtf8TextInput,
-  createArrowUtf8TextIndexAccessor,
-  decodeArrowUtf8CodePoints,
-  isArrowUtf8DictionaryType,
-  isArrowUtf8DictionaryVector,
-  isArrowUtf8TextVector,
-  isArrowUtf8Vector,
-  populateUtf8TextIndices,
-  type ArrowGlyphLayout,
-  type ArrowUtf8Dictionary,
-  type ArrowUtf8DictionaryIndexType,
-  type ArrowUtf8Chunk,
-  type ArrowUtf8TextAccessorContext,
-  type ArrowUtf8TextIndexAccessor,
-  type ArrowUtf8TextType,
-  type ArrowUtf8TextVector,
-  type Utf8TextIndexTarget
-} from './arrow-conversion/arrow-text';
-export {
   type GpuDictionaryCompressedTextStream,
   type GpuDictionaryUtf8TextInput,
   type GpuExpandedTextStream,
   type GpuUtf8TextInput,
-  type TextGlyphLayout,
-  type Utf8Dictionary,
-  type Utf8DictionaryIndexType,
-  type Utf8TextType,
-  type Utf8TextVector
+  type TextGlyphLayout
 } from './model-utils/gpu-text-types';
 export {
   createGpuDictionaryUtf8ExpandedInput,
@@ -86,45 +58,20 @@ export {
   type GpuUtf8MapShaderSourceOptions
 } from './model-utils/gpu-utf8-map';
 export {
-  DEFAULT_ARROW_TEXT_FS,
-  DEFAULT_ARROW_TEXT_SHADER_LAYOUT,
-  DEFAULT_ARROW_TEXT_VS,
-  DEFAULT_CLIPPED_ARROW_TEXT_SHADER_LAYOUT,
-  DEFAULT_CLIPPED_ARROW_TEXT_VS,
+  DEFAULT_TEXT_FS,
+  DEFAULT_TEXT_SHADER_LAYOUT,
+  DEFAULT_TEXT_VS,
+  DEFAULT_CLIPPED_TEXT_SHADER_LAYOUT,
+  DEFAULT_CLIPPED_TEXT_VS,
   DEFAULT_DICTIONARY_STORAGE_TEXT_SHADER_LAYOUT,
   DEFAULT_DICTIONARY_STORAGE_TEXT_SOURCE,
+  DEFAULT_ROW_INDEXED_STORAGE_TEXT_SHADER_LAYOUT,
+  DEFAULT_ROW_INDEXED_STORAGE_TEXT_SOURCE,
+  ROW_INDEXED_COMPACT_GLYPH_VERTEX_BYTE_STRIDE,
   DEFAULT_STORAGE_INDEXED_TEXT_SHADER_LAYOUT,
   DEFAULT_STORAGE_INDEXED_TEXT_SOURCE
 } from './model-utils/text-shaders';
-export {createArrowTextDefaultFragmentShaderUniforms} from './model-utils/text-fragment-uniforms';
-export {
-  buildArrowTextGlyphTable,
-  createArrowAttributeTextState,
-  createArrowDictionaryStorageTextState,
-  createArrowStorageTextState,
-  createStorageTextStateFromGPUVectors,
-  packStorageTextClipRects,
-  type ArrowAttributeTextInputProps,
-  type ArrowAttributeTextRenderProps,
-  type ArrowAttributeTextState,
-  type ArrowDictionaryStorageTextBatchState,
-  type ArrowDictionaryStorageTextInputProps,
-  type ArrowDictionaryStorageTextRenderBatchState,
-  type ArrowDictionaryStorageTextRenderProps,
-  type ArrowDictionaryStorageTextSourceVectors,
-  type ArrowDictionaryStorageTextState,
-  type ArrowStorageTextBatchState,
-  type ArrowStorageTextInputProps,
-  type ArrowStorageTextRenderBatchState,
-  type ArrowStorageTextRenderProps,
-  type ArrowStorageTextSourceVectors,
-  type ArrowStorageTextState,
-  type ArrowTextGlyphTable,
-  type ArrowTextRenderBatchState,
-  type ArrowTextSourceVectors,
-  type GPUVectorStorageTextBatch,
-  type GPUVectorStorageTextInputProps
-} from './arrow-conversion/convert-arrow-text-vectors';
+export {createTextDefaultFragmentShaderUniforms} from './model-utils/text-fragment-uniforms';
 export {
   AttributeTextModel,
   type AttributeTextModelProps,
@@ -152,25 +99,14 @@ export {
   type DictionaryTextState
 } from './models/dictionary-text-model';
 export {
+  assertAttributeTextGPUVectorInputs,
+  assertDictionaryTextGPUVectorInputs,
+  assertStorageTextGPUVectorInputs,
+  ATTRIBUTE_TEXT_GPU_INPUT_SCHEMA,
+  DICTIONARY_TEXT_GPU_INPUT_SCHEMA,
+  STORAGE_TEXT_GPU_INPUT_SCHEMA,
   type AttributeTextInputProps,
   type DictionaryTextInputProps,
-  type StorageTextInputProps
+  type StorageTextInputProps,
+  type TextInputProps
 } from './model-utils/text-model-props';
-export {
-  convertArrowTextToAttribute,
-  convertArrowTextToAttributeModelProps,
-  convertArrowTextToAttributeState,
-  type ArrowTextConversionColumns,
-  type ConvertedArrowTextData,
-  type ConvertArrowTextProps
-} from './arrow-conversion/convert-arrow-text-to-attribute';
-export {
-  convertArrowTextToStorage,
-  convertArrowTextToStorageModelProps,
-  convertArrowTextToStorageState
-} from './arrow-conversion/convert-arrow-text-to-storage';
-export {
-  convertArrowTextToDictionary,
-  convertArrowTextToDictionaryModelProps,
-  convertArrowTextToDictionaryState
-} from './arrow-conversion/convert-arrow-text-to-dictionary';
