+++ modules/text/src/text-2d/model-utils/text-model-props.ts
@@ -3,34 +3,185 @@
 // Copyright (c) vis.gl contributors
 
 import type {ModelProps} from '@luma.gl/engine';
-import type {GPUVector} from '@luma.gl/tables';
+import {
+  assertModelGPUVectorInputs,
+  type GPUVector,
+  type ModelGPUInputSchema,
+  type ModelGPUInputVectors,
+  type ValueList,
+  type VertexList
+} from '@luma.gl/tables';
 import type {FontAtlas, FontSettings} from '../atlas/font-atlas-manager';
 import type {CharacterMapping} from '../atlas/text-utils';
 
-/** GPUVector inputs for attribute-backed 2D text preparation. */
-export interface AttributeTextInputProps extends ModelProps {
+const TEXT_DICTIONARY_INDEX_FORMATS = [
+  'sint8',
+  'sint16',
+  'sint32',
+  'uint8',
+  'uint16',
+  'uint32'
+] as const;
+const TEXT_FORMATS = ['value-list<uint8>', ...TEXT_DICTIONARY_INDEX_FORMATS] as const;
+
+/** Prepared GPU inputs consumed by attribute-backed 2D text models. */
+export const ATTRIBUTE_TEXT_GPU_INPUT_SCHEMA = [
+  {
+    name: 'positions',
+    kind: 'positions',
+    required: true,
+    formats: ['float32x2'],
+    source: 'source-mappable'
+  },
+  {
+    name: 'texts',
+    kind: 'text',
+    required: true,
+    formats: TEXT_FORMATS,
+    source: 'source-mappable'
+  },
+  {
+    name: 'colors',
+    kind: 'colors',
+    required: false,
+    formats: ['unorm8x4', 'vertex-list<unorm8x4>'],
+    source: 'source-mappable'
+  },
+  {
+    name: 'angles',
+    kind: 'scalars',
+    required: false,
+    formats: ['float32'],
+    source: 'source-mappable'
+  },
+  {
+    name: 'sizes',
+    kind: 'scalars',
+    required: false,
+    formats: ['float32'],
+    source: 'source-mappable'
+  },
+  {
+    name: 'pixelOffsets',
+    kind: 'positions',
+    required: false,
+    formats: ['float32x2'],
+    source: 'source-mappable'
+  },
+  {
+    name: 'clipRects',
+    kind: 'positions',
+    required: false,
+    formats: ['sint16x4'],
+    source: 'source-mappable'
+  }
+] as const satisfies ModelGPUInputSchema;
+
+/** Prepared GPU inputs consumed by storage-backed 2D text models. */
+export const STORAGE_TEXT_GPU_INPUT_SCHEMA = [
+  {
+    name: 'positions',
+    kind: 'positions',
+    required: true,
+    formats: ['float32x2'],
+    source: 'source-mappable'
+  },
+  {
+    name: 'texts',
+    kind: 'text',
+    required: true,
+    formats: TEXT_FORMATS,
+    source: 'source-mappable'
+  },
+  {
+    name: 'colors',
+    kind: 'colors',
+    required: false,
+    formats: ['unorm8x4'],
+    source: 'source-mappable'
+  },
+  {
+    name: 'angles',
+    kind: 'scalars',
+    required: false,
+    formats: ['float32'],
+    source: 'source-mappable'
+  },
+  {
+    name: 'sizes',
+    kind: 'scalars',
+    required: false,
+    formats: ['float32'],
+    source: 'source-mappable'
+  },
+  {
+    name: 'pixelOffsets',
+    kind: 'positions',
+    required: false,
+    formats: ['float32x2'],
+    source: 'source-mappable'
+  },
+  {
+    name: 'textAnchors',
+    kind: 'scalars',
+    required: false,
+    formats: ['uint8'],
+    source: 'source-mappable'
+  },
+  {
+    name: 'alignmentBaselines',
+    kind: 'scalars',
+    required: false,
+    formats: ['uint8'],
+    source: 'source-mappable'
+  },
+  {
+    name: 'clipRects',
+    kind: 'positions',
+    required: false,
+    formats: ['sint16x4'],
+    source: 'source-mappable'
+  }
+] as const satisfies ModelGPUInputSchema;
+
+/** Prepared GPU inputs consumed by dictionary storage-backed 2D text models. */
+export const DICTIONARY_TEXT_GPU_INPUT_SCHEMA = [
+  {
+    name: 'positions',
+    kind: 'positions',
+    required: true,
+    formats: ['float32x2'],
+    source: 'source-mappable'
+  },
+  {
+    name: 'texts',
+    kind: 'text',
+    required: true,
+    formats: TEXT_DICTIONARY_INDEX_FORMATS,
+    source: 'source-mappable'
+  },
+  ...STORAGE_TEXT_GPU_INPUT_SCHEMA.slice(2)
+] as const satisfies ModelGPUInputSchema;
+
+/** GPUVector inputs shared by all 2D text model preparation paths. */
+export interface TextInputProps extends ModelProps {
   /** GPU-resident label origins aligned one-for-one with `texts`; each row is `[x, y]`. */
-  positions: GPUVector;
-  /** GPU UTF-8 or dictionary-encoded UTF-8 labels aligned row-for-row with `positions`. */
-  texts: GPUVector;
-  /**
-   * Optional GPU packed RGBA8 text colors aligned with label rows or label characters.
-   *
-   * Arrow's TypeScript type does not encode fixed-list length; conversion validates
-   * `FixedSizeList<Uint8>` rows have `listSize === 4`.
-   */
-  colors?: GPUVector;
+  positions: GPUVector<'float32x2'>;
+  /** GPU-resident UTF-8 value bytes or dictionary row keys aligned row-for-row with `positions`. */
+  texts: GPUVector<
+    ValueList<'uint8'> | 'sint8' | 'sint16' | 'sint32' | 'uint8' | 'uint16' | 'uint32'
+  >;
   /** Optional GPU per-row angles in degrees. */
-  angles?: GPUVector;
+  angles?: GPUVector<'float32'>;
   /** Optional GPU per-row deck-style text sizes. */
-  sizes?: GPUVector;
+  sizes?: GPUVector<'float32'>;
   /** Optional GPU per-row pixel offsets; each row is `[x, y]`. */
-  pixelOffsets?: GPUVector;
+  pixelOffsets?: GPUVector<'float32x2'>;
   /**
    * Optional GPU packed per-label clip rectangles `[x, y, width, height]`.
    * Negative width or height disables clipping on that axis.
    */
-  clipRects?: GPUVector;
+  clipRects?: GPUVector<'sint16x4'>;
   /** Character set for atlas generation. Pass `'auto'` when the adapter should derive it. */
   characterSet?: FontSettings['characterSet'] | 'auto';
   /** Font atlas generation settings. */
@@ -41,36 +192,24 @@ export interface AttributeTextInputProps extends ModelProps {
   characterMapping?: CharacterMapping;
   /** Optional prebuilt atlas for texture binding when `characterMapping` is injected. */
   fontAtlas?: FontAtlas;
+}
+
+/** GPUVector inputs for attribute-backed 2D text preparation. */
+export interface AttributeTextInputProps extends TextInputProps {
+  /** Optional GPU packed RGBA8 text colors aligned with label rows or label characters. */
+  colors?: GPUVector<'unorm8x4' | VertexList<'unorm8x4'>>;
 }
 
 /** GPUVector inputs for storage-backed 2D text preparation. */
-export interface StorageTextInputProps extends ModelProps {
-  /** GPU-resident label origins aligned one-for-one with `texts`; each row is `[x, y]`. */
-  positions: GPUVector;
-  /** GPU UTF-8 or dictionary-encoded UTF-8 labels aligned row-for-row with `positions`. */
-  texts: GPUVector;
-  /**
-   * Optional GPU packed RGBA8 text colors aligned with label rows.
-   *
-   * Arrow's TypeScript type does not encode fixed-list length; conversion validates
-   * `FixedSizeList<Uint8>` rows have `listSize === 4`.
-   */
-  colors?: GPUVector;
-  /** Optional GPU per-row angles in degrees. */
-  angles?: GPUVector;
-  /** Optional GPU per-row deck-style text sizes. */
-  sizes?: GPUVector;
-  /** Optional GPU per-row pixel offsets; each row is `[x, y]`. */
-  pixelOffsets?: GPUVector;
+export interface StorageTextInputProps extends Omit<TextInputProps, 'texts'> {
+  /** GPU-resident UTF-8 value bytes or dictionary row keys aligned row-for-row with `positions`. */
+  texts: TextInputProps['texts'];
+  /** Optional GPU packed RGBA8 text colors aligned with label rows. */
+  colors?: GPUVector<'unorm8x4'>;
   /** Optional GPU per-row text anchor enum: 0=start, 1=middle, 2=end. */
-  textAnchors?: GPUVector;
+  textAnchors?: GPUVector<'uint8'>;
   /** Optional GPU per-row alignment baseline enum: 0=center, 1=top, 2=bottom. */
-  alignmentBaselines?: GPUVector;
-  /**
-   * Optional GPU packed per-label clip rectangles `[x, y, width, height]`.
-   * Negative width or height disables clipping on that axis.
-   */
-  clipRects?: GPUVector;
+  alignmentBaselines?: GPUVector<'uint8'>;
   /** Constant fallback color used when `colors` is absent. */
   color?: [number, number, number, number];
   /** Constant fallback angle in degrees used when `angles` is absent. */
@@ -83,66 +222,48 @@ export interface StorageTextInputProps extends ModelProps {
   textAnchor?: 'start' | 'middle' | 'end';
   /** Constant fallback alignment baseline used when `alignmentBaselines` is absent. */
   alignmentBaseline?: 'center' | 'top' | 'bottom';
-  /** Character set for atlas generation. Pass `'auto'` when the adapter should derive it. */
-  characterSet?: FontSettings['characterSet'] | 'auto';
-  /** Font atlas generation settings. */
-  fontSettings?: FontSettings;
-  /** Multiplier applied to the atlas font size for one-line baseline layout. */
-  lineHeight?: number;
-  /** Optional deterministic mapping, mainly useful when atlas generation is managed externally. */
-  characterMapping?: CharacterMapping;
-  /** Optional prebuilt atlas for texture binding when `characterMapping` is injected. */
-  fontAtlas?: FontAtlas;
 }
 
 /** GPUVector inputs for compressed dictionary storage-backed 2D text preparation. */
-export interface DictionaryTextInputProps extends ModelProps {
-  /** GPU-resident label origins aligned one-for-one with `texts`; each row is `[x, y]`. */
-  positions: GPUVector;
-  /** GPU dictionary-encoded UTF-8 labels aligned row-for-row with `positions`. */
-  texts: GPUVector;
-  /**
-   * Optional GPU packed RGBA8 text colors aligned with label rows.
-   *
-   * Arrow's TypeScript type does not encode fixed-list length; conversion validates
-   * `FixedSizeList<Uint8>` rows have `listSize === 4`.
-   */
-  colors?: GPUVector;
-  /** Optional GPU per-row angles in degrees. */
-  angles?: GPUVector;
-  /** Optional GPU per-row deck-style text sizes. */
-  sizes?: GPUVector;
-  /** Optional GPU per-row pixel offsets; each row is `[x, y]`. */
-  pixelOffsets?: GPUVector;
-  /** Optional GPU per-row text anchor enum: 0=start, 1=middle, 2=end. */
-  textAnchors?: GPUVector;
-  /** Optional GPU per-row alignment baseline enum: 0=center, 1=top, 2=bottom. */
-  alignmentBaselines?: GPUVector;
-  /**
-   * Optional GPU packed per-label clip rectangles `[x, y, width, height]`.
-   * Negative width or height disables clipping on that axis.
-   */
-  clipRects?: GPUVector;
-  /** Constant fallback color used when `colors` is absent. */
-  color?: [number, number, number, number];
-  /** Constant fallback angle in degrees used when `angles` is absent. */
-  angle?: number;
-  /** Constant fallback deck-style text size used when `sizes` is absent. */
-  size?: number;
-  /** Constant fallback pixel offset used when `pixelOffsets` is absent. */
-  pixelOffset?: [number, number];
-  /** Constant fallback text anchor used when `textAnchors` is absent. */
-  textAnchor?: 'start' | 'middle' | 'end';
-  /** Constant fallback alignment baseline used when `alignmentBaselines` is absent. */
-  alignmentBaseline?: 'center' | 'top' | 'bottom';
-  /** Character set for atlas generation. Pass `'auto'` when the adapter should derive it. */
-  characterSet?: FontSettings['characterSet'] | 'auto';
-  /** Font atlas generation settings. */
-  fontSettings?: FontSettings;
-  /** Multiplier applied to the atlas font size for one-line baseline layout. */
-  lineHeight?: number;
-  /** Optional deterministic mapping, mainly useful when atlas generation is managed externally. */
-  characterMapping?: CharacterMapping;
-  /** Optional prebuilt atlas for texture binding when `characterMapping` is injected. */
-  fontAtlas?: FontAtlas;
+export interface DictionaryTextInputProps extends Omit<StorageTextInputProps, 'texts'> {
+  /** GPU-resident dictionary row keys aligned row-for-row with `positions`. */
+  texts: GPUVector<'sint8' | 'sint16' | 'sint32' | 'uint8' | 'uint16' | 'uint32'>;
+}
+
+/** Validates prepared attribute text GPU vectors before building render state. */
+export function assertAttributeTextGPUVectorInputs(props: AttributeTextInputProps): void {
+  assertTextGPUVectorInputs('AttributeTextModel', ATTRIBUTE_TEXT_GPU_INPUT_SCHEMA, props);
+}
+
+/** Validates prepared storage text GPU vectors before building render state. */
+export function assertStorageTextGPUVectorInputs(props: StorageTextInputProps): void {
+  assertTextGPUVectorInputs('StorageTextModel', STORAGE_TEXT_GPU_INPUT_SCHEMA, props);
+}
+
+/** Validates prepared dictionary text GPU vectors before building render state. */
+export function assertDictionaryTextGPUVectorInputs(props: DictionaryTextInputProps): void {
+  assertTextGPUVectorInputs('DictionaryTextModel', DICTIONARY_TEXT_GPU_INPUT_SCHEMA, props);
+}
+
+function assertTextGPUVectorInputs(
+  modelName: string,
+  schema: ModelGPUInputSchema,
+  props: TextInputProps
+): void {
+  const vectors = Object.fromEntries(
+    schema.map(input => [input.name, props[input.name as keyof TextInputProps]])
+  ) as ModelGPUInputVectors;
+  assertModelGPUVectorInputs(modelName, schema, vectors);
+  const rowCount = props.positions.length;
+  for (const input of schema) {
+    const vector = vectors[input.name];
+    if (!vector) {
+      continue;
+    }
+    if (vector.length !== rowCount) {
+      throw new Error(
+        `${modelName} ${input.name} rows must match positions rows (${vector.length} !== ${rowCount})`
+      );
+    }
+  }
 }
