+++ modules/text/tsconfig.json
@@ -8,7 +8,6 @@
     "outDir": "dist"
   },
   "references": [
-    {"path": "../arrow"},
     {"path": "../core"},
     {"path": "../engine"},
     {"path": "../tables"}
